refactor: プレイヤーキル時の処理をユースケースとして表現するように
差し替え可能にしたかったのでクラスにした
UI処理は別のクラスで処理する（未実装）
